refactor: Enable Featured tab in main GUI
The code changes enable the Featured tab in the main GUI by importing and adding the Featured class to the MainWindow class. This allows users to access and interact with the Featured tab, which was previously commented out. The changes also include the deletion of the unused main.py file in the gui/Featured directory.